class09/pyproject.toml
[project]
name = "class09"
version = "0.1.0"
description = ""
authors = [
    {name = "Hafiz-Zuhaib-Dev",email = "hafizzuhaibazaad457@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.5,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry.scripts]
dev='class09.main:start'

[tool.poetry]
packages = [{include = "class09", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
